+++ src/plugin.yml
@@ -1,5 +1,5 @@
 name: S.T.A.L.K.E.R
 main: ru.stalker.Main
 description: S.T.A.L.K.E.R
-version: 0.3
+version: 1.0
