--- src/ru/stalker/Bleeding.java
@@ -1,9 +1,0 @@
-package ru.stalker;
-
-import org.bukkit.entity.Player;
-
-public class Bleeding {
-	Player player;
-	boolean use;
-
-}
