+++ src/ru/stalker/Dimension.java
