--- stalker1/src/plugin.yml
@@ -1,5 +1,0 @@
-name: S.T.A.L.K.E.R
-main: ru.stalker.Main
-description: S.T.A.L.K.E.R
-version: 1.0
-
